requirements.txt
rocketmq-client-python==2.0.0
snowflake-id
